# protocol/openapi.yaml
openapi: 3.1.0
info:
  title: CapOwn Control Plane API
  version: 1.6.0
  description: |
    Language-independent v1.6 wire contract for the CapOwn control plane.

    Covers the connectivity milestone, plugin discovery, task dispatch,
    claim-based job delivery, results, cancellation, and plugin invocation
    via the `plugin_call` task type. Workers claim jobs via long-poll;
    SSE is an optional wake accelerator.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  x-capown-protocol-version: "1.6"
  x-capown-plugin-error-codes:
    - plugin_not_found
    - plugin_disabled
    - plugin_unavailable
    - plugin_tool_not_found
    - plugin_schema_invalid
    - plugin_timeout
    - plugin_canceled
    - plugin_output_too_large
    - plugin_protocol_error
    - plugin_concurrency_exceeded
  x-capown-task-error-codes:
    - worker_offline
  x-capown-auth-error-codes:
    - user_disabled
    - invitation_invalid
servers:
  - url: http://localhost:9210
    description: Local development Master
...
paths:
...
  /v1/workers/{worker_id}:
    parameters:
      - $ref: '#/components/parameters/WorkerId'
    get:
      tags: [workers]
      summary: Get one visible Worker
      operationId: getWorker
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web, client, admin, worker]
      responses:
        "200":
          description: Worker information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/WorkerNotFound'
    patch:
      tags: [workers]
      summary: Rename a Worker
      operationId: renameWorker
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameWorkerRequest'
      responses:
        "200":
          description: Worker renamed.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: updated
        "400":
          $ref: '#/components/responses/InvalidInput'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/WorkerNotFound'
        "409":
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [workers]
      summary: Revoke a Worker
      operationId: revokeWorker
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      responses:
        "204":
          description: Worker revoked.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/WorkerNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Send `Authorization: Bearer <token>`. Token classes are identified by
        server-side storage and, for web/worker sessions, by the cown_* prefix.
  parameters:
    WorkerId:
      name: worker_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/WorkerId'
...
  responses:
    InvalidInput:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Credentials are valid but do not grant the requested access.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    WorkerNotFound:
      description: Worker not found or revoked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          examples: [invalid_input]
        message:
          type: string
          description: Human-readable diagnostic message.
        details:
          description: Optional structured details. Current v1 commonly returns null.
          type: [object, array, string, number, boolean, 'null']
...
    Worker:
      type: object
      additionalProperties: false
      required: [worker_id, worker_name, owner_user_id, owner_username, hostname, os, mode, capabilities, workspace, status, last_heartbeat, registered_at, previous_worker_name, renamed_at, plugins]
      properties:
        worker_id:
          $ref: '#/components/schemas/WorkerId'
        worker_name:
          $ref: '#/components/schemas/WorkerName'
        owner_user_id:
          $ref: '#/components/schemas/ShortId'
        owner_username:
          type: string
          minLength: 1
        hostname:
          type: string
        os:
          type: string
        mode:
          type: string
        capabilities:
          type: array
          items:
            type: string
        workspace:
          type: string
        status:
          type: string
          enum: [online, offline]
        last_heartbeat:
          type: [string, 'null']
          format: date-time
        registered_at:
          type: [string, 'null']
          format: date-time
        previous_worker_name:
          type: [string, 'null']
        renamed_at:
          type: [string, 'null']
        plugins:
          type: array
          items:
            $ref: '#/components/schemas/PluginInfo'
          description: Plugin snapshot from the latest runtime update.
...
    RenameWorkerRequest:
      type: object
      additionalProperties: false
      required: [worker_name]
      properties:
        worker_name:
          $ref: '#/components/schemas/WorkerName'
...
    PluginToolInfo:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          default: ""
        input_schema:
          type: object
          default: {}
    PluginInfo:
      type: object
      additionalProperties: false
      required: [plugin_id, version, kind, transport, enabled, status, tools, error]
      properties:
        plugin_id:
          type: string
          description: Unique plugin identifier within this Worker.
        version:
          type: string
        kind:
          type: string
          description: Plugin kind, e.g. `mcp`.
        transport:
          type: string
          description: Plugin transport, e.g. `stdio`.
        enabled:
          type: boolean
          description: Persistent desired state from the Worker plugin manifest.
        status:
          type: string
          enum: [starting, running, stopped, error, disabled]
        tools:
          type: array
          items:
            $ref: '#/components/schemas/PluginToolInfo'
        error:
          type: string
          description: Sanitized error message when status is `error`.
...
    WorkerId:
      type: string
      pattern: '^wrk_[0-9a-f]{24}$'
    ShortId:
      type: string
      pattern: '^[0-9a-f]{12}$'
    WorkerName:
      type: string
      minLength: 3
      maxLength: 48
      pattern: '^[a-z0-9][a-z0-9._-]{1,46}[a-z0-9]$'
      description: Lowercase worker name; reserved names and wrk_ prefix are disallowed.
...